Initial Conversion to Forge 1.18.0 (38.0.17)
Everything works both on client and on a server! Yay!

Mainly just had to rename everything to Forge-equivalent mappings and add a few extra files for it to load properly. Will definitely be tweaking a few things here and there (mainly removing nonessentials).

Also pushed to a completely new branch.